bug: invalid conversion be->le
The members of the struct btrfs_ioctl_search_header are already in the
native endianess and these don't need a conversion little endian ->
native format. So drop the get_u32()/get_u64() call for these.

This is not a problem in the x86 world, which is little endian. But
in the (e.g.) ppc world, the native format is big endian so using a
get_u32()/get_u64() functions cause an incorrect conversion le -> be
when the values are already in the native format.

At the same time replace the get_u32()/get_u64() function with the
get_unaligned_[le]32/64() functions.

Signed-off-by: Goffredo Baroncelli <kreijack@libero.it>